+++ ChunkyMode/ChunkyOptionalMods.cs
@@ -2,6 +2,8 @@ using System.Runtime.CompilerServices;
 using RoR2;
 using MaterialHud;
 using UnityEngine;
+using Mono.Cecil.Cil;
+using MonoMod.Cil;
 
 namespace HDeMods { namespace ChunkyOptionalMods {
     internal static class Hunk {
@@ -19,5 +21,36 @@ namespace HDeMods { namespace ChunkyOptionalMods {
         [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
         public static void AddLegacyDifficulty() => RiskUIPlugin.DifficultyIconMap["CHUNKYMODEDIFFMOD_NAME"] = 
             ChunkyMode.HurricaneBundle.LoadAsset<Sprite>("texChunkyModeRiskUI");
+    }
+
+    internal static class Starstorm2 {
+        public static bool Enabled => BepInEx.Bootstrap.Chainloader.PluginInfos.ContainsKey("com.TeamMoonstorm");
+        private const float chirrHealOverride = 2f;
+        
+        [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
+        public static void SetHooks() => IL.SS2.Survivors.Chirr.ModifyStats += BuffChirr;
+        
+        [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
+        public static void RemoveHooks() => IL.SS2.Survivors.Chirr.ModifyStats -= BuffChirr;
+        
+        
+        [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
+        private static void BuffChirr(ILContext il) {
+            ILCursor c = new ILCursor(il);
+            if (!c.TryGotoNext(
+                    moveType: MoveType.After,
+                    x => x.MatchLdsfld("SS2.Survivors.Chirr", "_percentHealthRegen")
+                )) {
+                CM.Log.Error("Failed to hook Chirr Healing buff!");
+                return;
+            }
+            c.Emit(OpCodes.Ldarg_1);
+            c.EmitDelegate<RuntimeILReferenceBag.FastDelegateInvokers.Func<float, CharacterBody, float>>(
+                (toHeal, cb) => {
+                    if (cb.teamComponent.teamIndex != TeamIndex.Player) return toHeal;
+                    if (ChunkyMode.isSimulacrumRun && !ChunkyMode.waveStarted) return toHeal;
+                    return toHeal * chirrHealOverride;
+                });
+        }
     }
 } }